fixing bug transfer value di daftar pemain di halaman market. fixing bug di perhitungan other_expenses untuk perk. fixing bug di total coin yg selalu melebih angka poin. jdinya di Floor semua.

+++ web/app/Controller/AppController.php
@@ -109,7 +109,7 @@ class AppController extends Controller {
 				$this->userRank = @$point['Point']['rank'];
 
 				//get recent cash
-				$this->cash = intval($this->Game->getCash($this->userData['team']['id']));
+				$this->cash = floatval($this->Game->getCash($this->userData['team']['id']));
 				
 				//temporary patch.  remove it after next match.
 				if($this->userPoints==0){
